+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/spright-components
 
-This log was last generated on Tue, 07 May 2024 16:24:36 GMT and should not be manually modified.
+This log was last generated on Wed, 08 May 2024 17:29:00 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.18
+
+Wed, 08 May 2024 17:29:00 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v28.0.4
+
+## 0.0.17
+
+Tue, 07 May 2024 20:15:55 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v28.0.3
 
 ## 0.0.16